pdc: revised funcdtions for new abi event emission
